E-Commerce Dashboard.docx: Report, E-Commerce.pbix
This project presents the outcomes of a comprehensive data analysis conducted using Power BI. The primary goal of this project was to transform raw e-commerce sales data into actionable insights that support informed decision-making and enhance business performance. The analysis focused on various metrics, visualizations, and trends to provide a clear understanding of sales dynamics and customer preferences.